resourcegen: move Patch* to mutators.go
